refactor: Simplify join page with PincodePage component
- Replace complex join page implementations with a reusable PincodePage component
- Reduce code duplication between `/join` and `/join/[pincode]` routes
- Improve maintainability by centralizing pincode joining logic
